+++ lib/src/platform_date_picker.dart
@@ -10,7 +10,9 @@ import 'package:flutter/material.dart'
         showDatePicker,
         SelectableDayPredicate,
         DatePickerEntryMode,
-        DatePickerMode;
+        DatePickerMode,
+        CalendarDelegate,
+        GregorianCalendarDelegate;
 import 'package:flutter/widgets.dart';
 
 import 'platform.dart';
@@ -46,7 +48,7 @@ abstract class _BaseData {
   final DateTime? firstDate;
   final DateTime? lastDate;
 
-  const _BaseData({this.initialDate, this.firstDate, this.lastDate});
+  const   _BaseData({this.initialDate, this.firstDate, this.lastDate});
 }
 
 class MaterialDatePickerData extends _BaseData {
@@ -78,6 +80,7 @@ class MaterialDatePickerData extends _BaseData {
     this.barrierColor,
     this.barrierDismissible,
     this.barrierLabel,
+    this.calendarDelegate,
   });
 
   final DateTime? currentDate;
@@ -104,6 +107,7 @@ class MaterialDatePickerData extends _BaseData {
   final Color? barrierColor;
   final bool? barrierDismissible;
   final String? barrierLabel;
+  final CalendarDelegate<DateTime>? calendarDelegate;
 }
 
 class CupertinoDatePickerData extends _BaseData {
@@ -125,6 +129,7 @@ class CupertinoDatePickerData extends _BaseData {
     this.showDayOfWeek,
     this.itemExtent,
     this.selectionOverlayBuilder,
+    this.showTimeSeparator,
   });
 
   final Key? key;
@@ -141,6 +146,7 @@ class CupertinoDatePickerData extends _BaseData {
   final bool? showDayOfWeek;
   final double? itemExtent;
   final SelectionOverlayBuilder? selectionOverlayBuilder;
+  final bool? showTimeSeparator;
 }
 
 Future<DateTime?> showPlatformDatePicker({
@@ -183,6 +189,8 @@ Future<DateTime?> showPlatformDatePicker({
       barrierColor: data?.barrierColor,
       barrierDismissible: data?.barrierDismissible ?? true,
       barrierLabel: data?.barrierLabel,
+      calendarDelegate:
+          data?.calendarDelegate ?? const GregorianCalendarDelegate(),
     );
   } else {
     final data = cupertino?.call(context, platform(context));
@@ -284,6 +292,7 @@ class DefaultCupertinoDatePicker extends StatelessWidget {
             showDayOfWeek: data?.showDayOfWeek ?? false,
             itemExtent: data?.itemExtent ?? _kItemExtent,
             selectionOverlayBuilder: data?.selectionOverlayBuilder,
+            showTimeSeparator: data?.showTimeSeparator ?? false,
           ),
           Row(
             children: [