+++ lib/src/platform_text_button.dart
@@ -63,6 +63,10 @@ class CupertinoTextButtonData extends _BaseData {
     this.padding,
     this.disabledColor,
     this.borderRadius,
+    @Deprecated(
+      'Use minimumSize instead. '
+      'This feature was deprecated after v3.28.0-3.0.pre.',
+    )
     this.minSize,
     this.pressedOpacity,
     this.alignment,
@@ -73,6 +77,8 @@ class CupertinoTextButtonData extends _BaseData {
     this.onFocusChange,
     this.onLongPress,
     this.sizeStyle,
+    this.minimumSize,
+    this.mouseCursor,
   });
 
   final Color? color;
@@ -88,6 +94,8 @@ class CupertinoTextButtonData extends _BaseData {
   final ValueChanged<bool>? onFocusChange;
   final VoidCallback? onLongPress;
   final CupertinoButtonSize? sizeStyle;
+  final Size? minimumSize;
+  final MouseCursor? mouseCursor;
 
   // If true will use the filled style rather than the text style
   final bool originalStyle;
@@ -180,7 +188,8 @@ class PlatformTextButton extends PlatformWidgetBase<Widget, TextButton> {
         child: data?.child ?? child!,
         onPressed: data?.onPressed ?? onPressed,
         borderRadius:
-            data?.borderRadius ?? const BorderRadius.all(Radius.circular(8.0)),
+            data?.borderRadius ??
+            const BorderRadius.all(const Radius.circular(8.0)),
         minSize: data?.minSize ?? _kMinInteractiveDimensionCupertino,
         padding: data?.padding ?? padding,
         pressedOpacity: data?.pressedOpacity ?? 0.4,
@@ -193,6 +202,9 @@ class PlatformTextButton extends PlatformWidgetBase<Widget, TextButton> {
         onFocusChange: data?.onFocusChange,
         onLongPress: data?.onLongPress,
         sizeStyle: data?.sizeStyle ?? CupertinoButtonSize.large,
+        color: data?.color ?? color,
+        minimumSize: data?.minimumSize,
+        mouseCursor: data?.mouseCursor,
       );
       if (color != null) {
         final themeData = CupertinoTheme.of(context);
@@ -208,7 +220,8 @@ class PlatformTextButton extends PlatformWidgetBase<Widget, TextButton> {
         child: data?.child ?? child!,
         onPressed: data?.onPressed ?? onPressed,
         borderRadius:
-            data?.borderRadius ?? const BorderRadius.all(Radius.circular(8.0)),
+            data?.borderRadius ??
+            const BorderRadius.all(const Radius.circular(8.0)),
         minSize: data?.minSize ?? _kMinInteractiveDimensionCupertino,
         padding: data?.padding ?? padding,
         pressedOpacity: data?.pressedOpacity ?? 0.4,
@@ -222,6 +235,8 @@ class PlatformTextButton extends PlatformWidgetBase<Widget, TextButton> {
         onFocusChange: data?.onFocusChange,
         onLongPress: data?.onLongPress,
         sizeStyle: data?.sizeStyle ?? CupertinoButtonSize.large,
+        minimumSize: data?.minimumSize,
+        mouseCursor: data?.mouseCursor,
       );
     }
   }